fix: refactor headers
