+++ front/gatsby/src/components/Layout/Header/Subheaders/index.tsx
@@ -1,0 +1,15 @@
+// Copyright 2018-2020 @paritytech/Nomidot authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import BlockHeader from './BlockHeader'
+import EraHeader from './EraHeader';
+import SessionHeader from './SessionHeader';
+import StakingHeader from './StakingHeader';
+
+export {
+  BlockHeader,
+  EraHeader,
+  SessionHeader,
+  StakingHeader
+}
